Move attestation verification into internal/spaces
Client attestation JWT verification is domain logic, not HTTP
handling — moves it out of internal/pearserver into internal/spaces
alongside the rest of the space abstraction. Exports VerifyAttestation
and the AttestationTyp constant since it's now cross-package, and
relocates the AttestationTestClient/SignAttestation test helpers into
internal/spaces/testutil (a real importable package, unlike a _test.go
file) so both internal/spaces's own tests and
internal/pearserver's GetSpaceCredential tests can use them. Also
renames get_space_credential_test.go to match its handler file,
space_get_space_credential.go.
