+++ internal/pearserver/space_get_space_credential_test.go
@@ -11,6 +11,7 @@ import (
 	httpx_testutil "github.com/habitat-network/habitat/internal/httpx/testutil"
 	"github.com/habitat-network/habitat/internal/pearserver"
 	pearserver_testutil "github.com/habitat-network/habitat/internal/pearserver/testutil"
+	spaces_testutil "github.com/habitat-network/habitat/internal/spaces/testutil"
 )
 
 func TestServer_GetSpaceCredential_OpenSpaceNoAttestation(t *testing.T) {
@@ -35,7 +36,7 @@ func TestServer_GetSpaceCredential_AllowListSpace(t *testing.T) {
 
 	// GetSpaceCredential verifies the attestation's aud against the space
 	// owner DID, which org's CreateSpace call above makes owner==org.
-	clientID, priv := pearserver.AttestationTestClient(t, "key-1")
+	clientID, priv := spaces_testutil.AttestationTestClient(t, "key-1")
 	pearserver.GrantAppAccess(t, ts.SpaceStore, uri, clientID)
 
 	t.Run("no attestation is rejected", func(t *testing.T) {
@@ -50,7 +51,7 @@ func TestServer_GetSpaceCredential_AllowListSpace(t *testing.T) {
 	})
 
 	t.Run("granted client with valid attestation is accepted", func(t *testing.T) {
-		attestation := pearserver.SignAttestation(t, priv, "key-1", clientID, org, nil)
+		attestation := spaces_testutil.SignAttestation(t, priv, "key-1", clientID, org, nil)
 		var out habitat.NetworkHabitatSpaceGetSpaceCredentialOutput
 		code := httpx_testutil.NewTestXRPCClient(t).Procedure(
 			ts.Server.GetSpaceCredential,
@@ -64,8 +65,8 @@ func TestServer_GetSpaceCredential_AllowListSpace(t *testing.T) {
 	})
 
 	t.Run("non-granted client is rejected", func(t *testing.T) {
-		otherClientID, otherPriv := pearserver.AttestationTestClient(t, "key-1")
-		attestation := pearserver.SignAttestation(
+		otherClientID, otherPriv := spaces_testutil.AttestationTestClient(t, "key-1")
+		attestation := spaces_testutil.SignAttestation(
 			t,
 			otherPriv,
 			"key-1",
@@ -104,8 +105,8 @@ func TestServer_GetSpaceCredential_OpenSpaceWithValidAttestation(t *testing.T) {
 	uri, err := ts.SpaceStore.CreateSpace(t.Context(), org, groupTp, "test")
 	require.NoError(t, err)
 
-	clientID, priv := pearserver.AttestationTestClient(t, "key-1")
-	attestation := pearserver.SignAttestation(t, priv, "key-1", clientID, org, nil)
+	clientID, priv := spaces_testutil.AttestationTestClient(t, "key-1")
+	attestation := spaces_testutil.SignAttestation(t, priv, "key-1", clientID, org, nil)
 
 	var out habitat.NetworkHabitatSpaceGetSpaceCredentialOutput
 	code := httpx_testutil.NewTestXRPCClient(t).Procedure(
